+++ packages/sdk/src/pub/plt/TokenHolder.ts
@@ -1,1 +1,11 @@
-export { Type, fromJSON, instanceOf, fromAccountAddress, JSON, Account, AccountJSON } from '../../plt/TokenHolder.js';
+export {
+    Type,
+    fromJSON,
+    instanceOf,
+    fromAccountAddress,
+    JSON,
+    Account,
+    AccountJSON,
+    fromBase58,
+    toBase58,
+} from '../../plt/TokenHolder.js';